+++ src/main/kotlin/com/cmgapps/license/reporter/Report.kt
@@ -47,47 +47,7 @@ abstract class LicensesSingleFileReport(
 
     override fun getDisplayName(): String = "License Report for ${type.name}"
 
-    @Override
-    @Deprecated("Deprecated in Java", replaceWith = ReplaceWith("getOutputLocation().set"))
-    override fun setDestination(file: File) {
-        outputLocation.fileValue(file)
-    }
-
     override fun getOutputType(): Report.OutputType = Report.OutputType.FILE
-
-    /**
-     * Needed for 7.x
-     */
-    @Override
-    fun setDestination(provider: Provider<File>) {
-        outputLocation.fileProvider(provider)
-    }
-
-    /**
-     * Needed for 7.x
-     */
-    @Override
-    fun getDestination(): File = outputLocation.get().asFile
-
-    /**
-     * Needed for 7.x
-     */
-    @Override
-    fun isEnabled(): Boolean = required.get()
-
-    /**
-     * Needed for 7.x
-     */
-    @Override
-    fun setEnabled(enabled: Boolean) {
-        required.set(enabled)
-    }
-
-    /**
-     * Needed for 7.x
-     */
-    @Override
-    fun setEnabled(enabled: Provider<Boolean>) = required.set(enabled)
 
     override fun configure(cl: Closure<*>): Report {
         cl.call(this)